pyproject.toml
[project]
name = "chartmogul-mcp-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chartmogul>=4.6.0",
    "mcp[cli]>=1.7.1",
    "python-dotenv>=1.1.0",
]
